Securized routes based on the User's role (if authenticated)

+++ app/Http/Controllers/ArtistController.php
@@ -4,20 +4,36 @@ namespace App\Http\Controllers;
 
 use App\Artist;
 use App\Http\Resources\ArtistResource;
-use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 
 class ArtistController extends Controller
 {
+    const PERMISSION_ERROR = 'You are not the manager of this artist';
+
+    /**
+     * Create a new ArtistController instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware(['auth:api', 'auth.manager:api'], ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
+     * @param Request $request
      * @return \Illuminate\Http\Resources\Json\AnonymousResourceCollection|\Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        return ArtistResource::collection(Artist::all());
+        $user = Auth::user();
+        if ($user && $user->role === "promoter" && $request->has('only_mine')) {
+            return ArtistResource::collection(Artist::whereManagerId($user->id)->get());
+        } else return ArtistResource::collection(Artist::all());
     }
 
     /**
@@ -30,8 +46,8 @@ class ArtistController extends Controller
     {
         // Metemos en la petición nuevos datos a calzador que generamos nosotros mismos de forma automática
         $request->merge([
-            'permalink' => kebab_case($request['name']),
-            'manager_id' => User::whereRole('manager')->first()->id // FIXME Obtener desde el Auth el ID del manager
+            'permalink' => kebab_case($request->name),
+            'manager_id' => Auth::user()->id // El middleware se asegura de que tenga rol "manager"
         ]);
 
         $validatedData = $request->validate([
@@ -54,7 +70,7 @@ class ArtistController extends Controller
      */
     public function show(Artist $artist)
     {
-        // Usamos el load para forzar el eager loading de los festivales
+        // Usamos el load para forzar el eager loading
         return ArtistResource::make($artist->load(['festivals', 'genres']));
     }
 
@@ -67,9 +83,13 @@ class ArtistController extends Controller
      */
     public function update(Request $request, Artist $artist)
     {
+        if (!$this->userOwns($artist)) {
+            return response()->json(['message' => self::PERMISSION_ERROR], Response::HTTP_UNAUTHORIZED);
+        }
+
         // Metemos en la petición nuevos datos a calzador que generamos nosotros mismos de forma automática
         $request->merge([
-            'permalink' => kebab_case($request['name']),
+            'permalink' => kebab_case($request->name),
         ]);
         // Decidimos si tenemos que buscar si existe el nuevo permalink o no
         // en función de si este ha cambiado respecto al original
@@ -93,11 +113,24 @@ class ArtistController extends Controller
      */
     public function destroy(Artist $artist)
     {
+        if (!$this->userOwns($artist)) {
+            return response()->json(['message' => self::PERMISSION_ERROR], Response::HTTP_UNAUTHORIZED);
+        }
+
         try {
             $artist->delete();
             return response(null, Response::HTTP_NO_CONTENT);
         } catch (\Exception $e) {
             return response()->json(['message' => $e->getMessage()], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
+    }
+
+    /**
+     * @param Artist $artist
+     * @return bool
+     */
+    private function userOwns(Artist $artist): bool
+    {
+        return Auth::user()->id === $artist->manager_id;
     }
 }
